start of proc gen

+++ src/components/win_state.rs
@@ -113,7 +113,7 @@ pub struct GameState {
     ///Current Game Win State
     pub ws: GameStateEnum,
     ///The level for which *ws* refers to
-    pub level_from: Either<usize, f32>,
+    pub level_from: Either<usize, u32>,
     ///Amount of time the level has taken
     pub level_no_of_moves: i32,
 }
@@ -134,7 +134,7 @@ impl GameState {
     ///  - **level_timer_len** refers to how long the level took
     pub fn new(
         lost_tile: Option<Option<TileTransform>>,
-        level_from: Either<usize, f32>,
+        level_from: Either<usize, u32>,
         level_timer_len: i32,
     ) -> Self {
         match lost_tile {
